chore: Upgrade csv-safe to the latest version (#9739)
The following error starting is shown on the console after the ruby
upgrade.

csv.rb was loaded from the standard library, but will no longer be part
of the default gems since Ruby 3.4.0. Add csv to your Gemfile or
gemspec. Also contact author of csv-safe-3.2.1 to add csv into its
gemspec.


Csv-safe has already added a patch via
https://github.com/zvory/csv-safe/pull/20.

This PR updates the version to the latest version of csv-safe (3.3.1)